requirements.txt
matplotlib
numpy
scipy
torch
neuraloperator
wandb
torchinfo
